+++ src/main/java/com/restaurante01/api_restaurante/modulos/produto/api/dto/entrada/ProdutoDTO.java
@@ -22,7 +22,7 @@ public record ProdutoDTO(
 
         @NotNull(message = "Quantidade deve ser zero ou positivo")
         @Min(value = 0, message = "Quantidade minima deve ser zero")
-        Long quantidadeAtual,
+        Integer quantidadeAtual,
 
         Boolean disponibilidade,
